improve: file structure
